Homegear quits when ntp-sync takes too long
I have installed homegear on a RPi. After a power loss it takes some time for my router (which is the NTP server in my network) to re-establish the internet connection and to synchronize its time with external servers (sometimes more than 15 minutes). This delay causes homegear to exit. In my opinion, there is no need for homgear to exit when the sync takes too long. By this change, homegear will wait endlessly until the ntp sync is completed.